refactor: 스트릭 컴포넌트 구조 리팩토링
