+++ src/components/heatmap/calendar/index.ts
@@ -1,0 +1,2 @@
+export * from './calendar-heatmap'
+export type * from './types'
