+++ src/components/heatmap/calendar/utils/cell-color.ts
