+++ src/components/heatmap/index.ts
@@ -1,1 +1,1 @@
-export * from './calendar/calendar-heatmap'
+export * from './calendar'
